LayoutTests/js/array-index-immediate-types-expected.txt

This test checks the behaviour of indexing an Array with immediate types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS array[0] is "Zero"
PASS array[null] is undefined
PASS array[undefined] is undefined
PASS array[true] is undefined
PASS array[false] is undefined
PASS putSelf([0], 0); is true
PASS putSelf([0], 1/9); is true
PASS successfullyParsed is true

TEST COMPLETE
